requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.3.0=mkl
absl-py=0.7.0=py36_0
appnope=0.1.0=py36hf537a9a_0
astor=0.7.1=py36_0
audioread=2.1.6=py36_0
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.1.0=py36_0
bzip2=1.0.6=h1de35cc_5
c-ares=1.15.0=h1de35cc_1
ca-certificates=2019.1.23=0
certifi=2018.11.29=py36_0
click=7.0=py36_0
cycler=0.10.0=py36hfc81398_0
dbus=1.13.6=h90a0687_0
decorator=4.3.2=py36_0
entrypoints=0.3=py36_0
expat=2.2.6=h0a44026_0
ffmpeg=4.0=h01ea3c9_0
freetype=2.9.1=hb4e5f40_0
gast=0.2.2=py36_0
gettext=0.19.8.1=h15daf44_3
glib=2.56.2=hd9629dc_0
grpcio=1.16.1=py36h044775b_1
h5py=2.9.0=py36h3134771_0
hdf5=1.10.4=hfa1e0ec_0
icu=58.2=h4b95b61_1
intel-openmp=2019.1=144
ipykernel=5.1.0=py36h39e3cac_0
ipython=7.2.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36h241746c_0
ipywidgets=7.4.2=py36_0
jedi=0.13.2=py36_0
jinja2=2.10=py36_0
joblib=0.13.1=py36_0
jpeg=9b=he5867d9_2
jsonschema=2.6.0=py36hb385e00_0
jupyter=1.0.0=py36_7
jupyter_client=5.2.4=py36_0
jupyter_console=6.0.0=py36_0
jupyter_core=4.4.0=py36_0
keras-applications=1.0.6=py36_0
keras-preprocessing=1.0.5=py36_0
kiwisolver=1.0.1=py36h0a44026_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libiconv=1.15=hdd342a3_7
libopus=1.3=h1de35cc_0
libpng=1.6.36=ha441bb4_0
libprotobuf=3.6.1=hd9629dc_0
librosa=0.6.2=py_0
libsodium=1.0.16=h3efe00b_0
libvpx=1.7.0=h378b8a2_0
llvmlite=0.27.0=py36h8c7ce04_0
markdown=3.0.1=py36_0
markupsafe=1.1.0=py36h1de35cc_0
matplotlib=3.0.2=py36h54f8f79_0
mistune=0.8.4=py36h1de35cc_0
mkl=2019.1=144
mkl_fft=1.0.10=py36h5e564d8_0
mkl_random=1.0.2=py36h27c97d8_0
nbconvert=5.3.1=py36_0
nbformat=4.4.0=py36h827af21_0
ncurses=6.1=h0a44026_1
notebook=5.7.4=py36_0
numba=0.42.0=py36h6440ff4_0
numpy=1.15.4=py36hacdab7b_0
numpy-base=1.15.4=py36h6575580_0
openssl=1.1.1a=h1de35cc_0
pandas=0.24.1=py36h0a44026_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parso=0.3.2=py36_0
pcre=8.42=h378b8a2_0
pexpect=4.6.0=py36_0
pickleshare=0.7.5=py36_0
pip=19.0.1=py36_0
prometheus_client=0.5.0=py36_0
prompt_toolkit=2.0.8=py_0
protobuf=3.6.1=py36h0a44026_0
ptyprocess=0.6.0=py36_0
pygments=2.3.1=py36_0
pyparsing=2.3.1=py36_0
pyqt=5.9.2=py36h655552a_2
python=3.6.8=haf84260_0
python-dateutil=2.7.5=py36_0
python-dotenv=0.10.1=py_0
pytz=2018.9=py36_0
pyzmq=17.1.2=py36h0a44026_2
qt=5.9.7=h468cd18_1
qtconsole=4.4.3=py36_0
readline=7.0=h1de35cc_5
resampy=0.2.1=py_1
scikit-learn=0.20.2=py36h27c97d8_0
scipy=1.2.0=py36h1410ff5_0
send2trash=1.5.0=py36_0
setuptools=40.7.3=py36_0
sip=4.19.8=py36h0a44026_0
six=1.12.0=py36_0
sqlite=3.26.0=ha441bb4_0
tensorboard=1.12.2=py36haf313ee_0
tensorflow=1.12.0=mkl_py36h2b2bbaf_0
tensorflow-base=1.12.0=mkl_py36h70e0e9a_0
termcolor=1.1.0=py36_1
terminado=0.8.1=py36_1
testpath=0.4.2=py36_0
tk=8.6.8=ha441bb4_0
tornado=5.1.1=py36h1de35cc_0
tqdm=4.29.1=py_0
traitlets=4.3.2=py36h65bd3ce_0
wcwidth=0.1.7=py36h8c6ec74_0
webencodings=0.5.1=py36_1
werkzeug=0.14.1=py36_0
wheel=0.32.3=py36_0
widgetsnbextension=3.4.2=py36_0
xz=5.2.4=h1de35cc_4
zeromq=4.3.1=h0a44026_3
zlib=1.2.11=h1de35cc_3
